major refactor into separate files for convenience
